requirements.txt
# langchain~=0.0.352
# psycopg2-binary~=2.9.9
# Unidecode~=1.3.7
# langchain-openai~=0.0.5
# flask
# python-dotenv
langsmith
langchain-community
langchain-core
langchain-experimental
langchain
langchain-openai
langserve
langchain-google-genai
google-generativeai
langchain-openai
langserve
langsmith
pillow
grpcio
psycopg2-binary
Unidecode
flask
python-dotenv
faiss-cpu
fastapi
uvicorn
boto3
sse_starlette
